requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

gspread==6.2.1
gspread-dataframe==4.0.0
gspread-formatting==1.2.1
numpy==2.2.6
pandas==2.2.3
psycopg2==2.9.10
PyAutoGUI==0.9.54
pyperclip==1.9.0
python-dotenv==1.1.0
pywin32==310
win32compat==221.26
pyautogui==0.9.54